New Account refactoring. Starts with terms and then proceeds to a combined about / account form page.
Changed the color of the right pane because it clashed with the header
color of the add account page. (you may not like it but it has to be
different than the old gray color, either that or we need to change the
color of the entry header).

Finally made the logo a link back to the entry page.